requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
docstring_parser==0.17.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.21
langchain-community==0.3.30
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langsmith==0.4.32
Markdown==3.9
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.3
propcache==0.4.0
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
zstandard==0.25.0
